Various
* now uses boost uuid for p2p
* replaces all size_t's for uint64_t (https://stackoverflow.com/a/28685755) - if something needs more than 64 bytes then we can change the call to uint128 and include this lib https://github.com/calccrypto/uint128_t
* adds c++17 support
* change a few things to bring it up to c++17
* working version proof: https://github.com/Lithe-Project/Lithe/tree/c++17 includes a modern c++ lib which complies and works
***
* it syncs to the network

+++ external/google/sparsehash/densehashtable.h
@@ -926,7 +926,7 @@ class dense_hashtable {
   // (1) Iterator supports operator-, resize before inserting
   template <class ForwardIterator>
   void insert(ForwardIterator f, ForwardIterator l, STL_NAMESPACE::forward_iterator_tag) {
-    size_t dist = STL_NAMESPACE::distance(f, l);
+    uint64_t dist = STL_NAMESPACE::distance(f, l);
     if (dist >= (std::numeric_limits<size_type>::max)())
       throw std::length_error("insert-range overflow");
     resize_delta(static_cast<size_type>(dist));
