+++ external/gtest/samples/sample2.cc
@@ -39,7 +39,7 @@
 const char* MyString::CloneCString(const char* a_c_string) {
   if (a_c_string == NULL) return NULL;
 
-  const size_t len = strlen(a_c_string);
+  const uint64_t len = strlen(a_c_string);
   char* const clone = new char[ len + 1 ];
   memcpy(clone, a_c_string, len + 1);
