+++ external/gtest/xcode/Samples/FrameworkSample/widget.h
@@ -50,7 +50,7 @@ class Widget {
 
   // Public accessors to the string data
   std::string GetStringValue() const;
-  void GetCharPtrValue(char* buffer, size_t max_size) const;
+  void GetCharPtrValue(char* buffer, uint64_t max_size) const;
 
  private:
   // Data members
